Backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2.3
django-dotenv==1.4.2
django-environ==0.10.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
